+ Added an export system to allow other modules to hook Squirrel + Rewrote part of the module core to prevent crashes + Added an OutputDebug method * Changes to some functions due to "performance" warnings when compiling in VS2010 * Rearranged a few files to go into different filters

+++ CVehicle.h
@@ -51,7 +51,7 @@ class CVehicle
 		bool GetStreamedForPlayer( CPlayer * player );
 		CPlayer * GetOccupant( int slot );
 		int SetHandlingData( int rule, float value );
-		float GetHandlingData( int rule );
+		double GetHandlingData( int rule );
 		void ResetHandlingData( int rule );
 		void ResetAllHandling();
 		bool IsHandlingSet( int rule );
